requirements/base.txt
astroid==1.6.2
autopep8==1.3.5
base58==1.0.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.4
django-environ==0.4.4
django-filter==1.1.0
djangorestframework==3.8.1
djangorestframework-simplejwt==3.2.3
ecdsa==0.13
graphviz==0.8.1
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.6.1
pylint==1.8.3
pytz==2018.3
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.11
